\name{AHPubMedDbs-package}
\alias{AHPubMedDbs-package}
\alias{AHPubMedDbs}
\docType{package}
\title{
\packageTitle{AHPubMedDbs}
}
\description{
\packageDescription{AHPubMedDbs}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AHPubMedDbs}
\packageIndices{AHPubMedDbs}
}
\author{
\packageAuthor{AHPubMedDbs}

Maintainer: \packageMaintainer{AHPubMedDbs}
}

\keyword{ package }
\examples{
ls("package:AHPubMedDbs") # This is empty package
}